rename folder, code refactoring, README
