+++ python/codepipeline-build-deploy-github/README.md
@@ -14,7 +14,7 @@
 
 ## Overview
 
-This AWS Cloud Development Kit (CDK) Python example demonstrates how to configure AWS CodePipeline with Github, CodeBuild, and CodeDeploy to build and deploy a Docker image to an Elastic Container Service (ECS) cluster running [AWS Fargate](https://aws.amazon.com/fargate/) (serverless compute for containers).
+This AWS Cloud Development Kit (CDK) Python example demonstrates how to configure AWS CodePipeline with GitHub, CodeBuild, and CodeDeploy to build and deploy a Docker image to an Elastic Container Service (ECS) cluster running [AWS Fargate](https://aws.amazon.com/fargate/) (serverless compute for containers).
 
 ## Real-world Example
 
@@ -25,6 +25,15 @@ When working in fast-paced development environments, CI/CD (Continuous Integrati
 - [Python v3.6+](https://www.python.org/)
   - [AWS CDK in Python](https://docs.aws.amazon.com/cdk/v2/guide/work-with-cdk-python.html)
 - [AWS CDK v2.x](https://docs.aws.amazon.com/cdk/v2/guide/getting_started.html)
+- GitHub Account
+  - Create [Personal Access Token](https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens), make sure to select the `admin:repo_hook` and `repo` permissions to allow read access to the pipeline and also allow it to install a webhook to trigger pipeline actions when making a new commit.
+  - Store PAT to Secrets Manager
+    ```
+    GITHUB_ACCESS_TOKEN='your access token'
+    REGION=us-east-1
+    aws secretsmanager  create-secret --name github_access_token --description "GitHub Access Token" --secret-string $GITHUB_ACCESS_TOKEN --region $REGION
+    ```
+
 
 ## AWS Services Utilized
 
@@ -39,8 +48,9 @@ When working in fast-paced development environments, CI/CD (Continuous Integrati
 ## Deploying
 
 - Authenticate to an AWS account via a Command Line Interface (CLI).
-- Navigate to this `codepipeline-build-deploy` directory.
+- Navigate to this `codepipeline-build-deploy-github` directory.
 - `pip install -r requirements.txt` to install the required dependencies.
+- Replace `github_owner_name` with your github [username](https://docs.github.com/en/account-and-profile/setting-up-and-managing-your-personal-account-on-github/managing-email-preferences/remembering-your-github-username-or-email#finding-your-username-in-the-url-of-remote-repositories) in `codepipeline_build_deploy_stack.py`
 - `cdk synth` to generate and review the CloudFormation template.
 - `cdk diff` to compare local changes with what is currently deployed.
 - `pytest` to run the unit tests we specify in `tests/unit/test_codepipeline_build_deploy_stack.py`.
@@ -66,7 +76,6 @@ After a successful deployment, CDK will output a public endpoint for:
 - Navigate to the AWS CodePipeline console and select `ImageBuildDeployPipeline`. Then click on `Release change` to trigger the pipeline and observe the workflow in action end-to-end.
 - Navigate to the AWS Console to view the services that were deployed:
   - CodePipeline pipeline
-  - Github repository
   - CodeBuild project
   - CodeDeploy application
   - ECS cluster
@@ -82,14 +91,3 @@ After a successful deployment, CDK will output a public endpoint for:
 - Protect the public Application Load Balancer (ALB) from attacks and malicious traffic using [Web Application Firewall (WAF)](https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html).
 - [Enable SSL/TLS](https://docs.aws.amazon.com/elasticloadbalancing/latest/application/create-https-listener.html) at the Application Load Balancer (ALB) using AWS Certificate Manager.
 - Set up a [Route 53 domain](https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/routing-to-elb-load-balancer.html) to route traffic to an Application Load Balancer (ALB).
-
-## Github connection
-- Prepare Github repository and create [Personal Access Token](https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/managing-your-personal-access-tokens)
-  - Make sure to select the `admin:repo_hook` and `repo` permissions to allow read access to the pipeline and also allow it to install a webhook to trigger pipeline actions when making a new commit.***
-
-- Store PAT to Secrets Manager
-  - `GITHUB_ACCESS_TOKEN='your access token'`
-  - `REGION=us-east-1`
-  - `aws secretsmanager  create-secret --name github_access_token --description "Github access token" --secret-string $GITHUB_ACCESS_TOKEN --region $REGION`
-
-
