+++ python/codepipeline-build-deploy-github/app.py
@@ -7,7 +7,7 @@ from codepipeline_build_deploy.codepipeline_build_deploy_stack import Codepipeli
 
 
 app = cdk.App()
-CodepipelineBuildDeployStack(app, "CodepipelineBuildDeployStackGithubPython",
+CodepipelineBuildDeployStack(app, "CodepipelineGithubBuildDeployStack",
     # If you don't specify 'env', this stack will be environment-agnostic.
     # Account/Region-dependent features and context lookups will not work,
     # but a single synthesized template can be deployed anywhere.
