+++ python/codepipeline-build-deploy-github/lambda/github_repo_creator/app/site-assets/index.html
