+++ python/codepipeline-build-deploy-github/requirements.txt
@@ -1,4 +1,3 @@
 aws-cdk-lib==2.154.1
 constructs>=10.0.0,<11.0.0
-requests=2.32.3
